Move all MongoDB Panache common classes to the common root package
This is a leftover from the housekeeping done on split packages
