+++ extensions/panache/mongodb-panache-common/runtime/src/main/java/io/quarkus/mongodb/panache/common/binder/PanacheQlQueryBinder.java
@@ -1,4 +1,4 @@
-package io.quarkus.mongodb.panache.binder;
+package io.quarkus.mongodb.panache.common.binder;
 
 import java.util.HashMap;
 import java.util.Map;
