+++ extensions/panache/mongodb-panache-common/runtime/src/main/java/io/quarkus/mongodb/panache/common/jsonb/ObjectIdDeserializer.java
@@ -1,4 +1,4 @@
-package io.quarkus.mongodb.panache.jsonb;
+package io.quarkus.mongodb.panache.common.jsonb;
 
 import java.lang.reflect.Type;
