+++ extensions/panache/mongodb-panache-common/runtime/src/main/java/io/quarkus/mongodb/panache/common/transaction/MongoTransactionException.java
@@ -1,4 +1,4 @@
-package io.quarkus.mongodb.panache.transaction;
+package io.quarkus.mongodb.panache.common.transaction;
 
 public class MongoTransactionException extends RuntimeException {
     public MongoTransactionException(Exception cause) {
